+++ Resources/tetris-bar-32x8.tsx
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<tileset name="bar-32x8" tilewidth="32" tileheight="8" tilecount="9" columns="1">
+ <image source="bar-32x8.png" trans="800080" width="32" height="72"/>
+</tileset>