+++ Resources/tetris-tshape-24x16.tsx
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<tileset name="tshape-24x16" tilewidth="24" tileheight="16" tilecount="9" columns="3">
+ <image source="tshape-24x16.png" trans="800080" width="72" height="48"/>
+ <tile id="8">
+  <objectgroup draworder="index">
+   <object id="1" x="0" y="16">
+    <polygon points="0,0 0,-8 8,-8 8,-16 16,-16 16,-7 24,-7 24,0"/>
+   </object>
+  </objectgroup>
+ </tile>
+</tileset>