+++ macOS/GameViewController.swift
@@ -40,6 +40,7 @@ class GameViewController: NSViewController {
         
         /* Sprite Kit applies additional optimizations to improve rendering performance */
         skView.ignoresSiblingOrder = true
+        skView.showsPhysics = false
         setupDebuggingLabels()